Finish feature to bin level 2 files when provided a list of files
Created a function to generate a .txt file containing the filenames of
the outputs of l2bin. In addition, added logic to retrieve the default
flags for modis. The process function should now be able to generate a
finished level 3 binned file when provided a list of files. However, it
currently does not remove any of the intermediate files such as the file
list .txt file.